refactor(app): drop blurred glass effect
